+++ components/script/layout_dom/servo_layout_node.rs
@@ -103,6 +103,12 @@ impl<'dom> ServoLayoutNode<'dom> {
     pub(super) unsafe fn dangerous_previous_sibling(&self) -> Option<Self> {
         self.node.prev_sibling_ref().map(Into::into)
     }
+
+    /// Whether this node or one of its ancestors has the `containertiming` attribute set.
+    /// <https://wicg.github.io/container-timing/>
+    pub fn has_container_timing(&self) -> bool {
+        unsafe { self.node.get_flag(NodeFlags::HAS_CONTAINER_TIMING) }
+    }
 }
 
 impl<'dom> From<LayoutDom<'dom, Node>> for ServoLayoutNode<'dom> {
@@ -340,4 +346,69 @@ impl NodeInfo for ServoLayoutNode<'_> {
     fn is_text_node(&self) -> bool {
         self.node.is_text_node_for_layout()
     }
+}
+
+/// Reconstruct a [`ServoLayoutNode`] from an [`style::dom::OpaqueNode`] obtained during
+/// display-list building.
+///
+/// # Safety
+///
+/// `opaque.0` must be a `JSObject` pointer for a live DOM node. Layout and display-list
+/// building run synchronously on the layout thread while the script thread is paused, so
+/// the GC cannot collect these objects in the meantime.
+unsafe fn servo_layout_node_from_opaque(opaque: style::dom::OpaqueNode) -> ServoLayoutNode<'static> {
+    use script_bindings::conversions::private_from_object;
+
+    unsafe {
+        let obj = opaque.0 as *mut js::jsapi::JSObject;
+        let raw = private_from_object(obj) as *const std::ffi::c_void;
+        ServoLayoutNode::new(&TrustedNodeAddress(raw))
+    }
+}
+
+/// Walk DOM ancestors of `opaque` to find the nearest element with a `containertiming`
+/// attribute and return that element's node.
+/// Called during display-list building from `check_for_container_timing_candidate`.
+/// <https://wicg.github.io/container-timing/>
+pub fn container_timing_root_for_node(
+    opaque: style::dom::OpaqueNode,
+) -> Option<style::dom::OpaqueNode> {
+    use html5ever::{LocalName, ns};
+
+    // Safety: see `servo_layout_node_from_opaque`.
+    let node = unsafe { servo_layout_node_from_opaque(opaque) };
+
+    let mut ancestor = unsafe { node.dangerous_dom_parent() };
+    loop {
+        match ancestor {
+            None => break None,
+            Some(current) => {
+                if let Some(element) = current.as_html_element() {
+                    if element
+                        .attribute(&ns!(), &LocalName::from("containertiming"))
+                        .is_some()
+                    {
+                        break Some(current.opaque());
+                    }
+                }
+                ancestor = unsafe { current.dangerous_dom_parent() };
+            },
+        }
+    }
+}
+
+/// Read the `containertiming` attribute value directly from `opaque`, which must already
+/// be a container timing root element.
+/// <https://wicg.github.io/container-timing/>
+pub fn container_timing_identifier_for_root(
+    opaque: style::dom::OpaqueNode,
+) -> Option<std::sync::Arc<str>> {
+    use html5ever::{LocalName, ns};
+
+    // Safety: see `servo_layout_node_from_opaque`.
+    let node = unsafe { servo_layout_node_from_opaque(opaque) };
+
+    let element = node.as_html_element()?;
+    let id = element.attribute(&ns!(), &LocalName::from("containertiming"))?;
+    Some(std::sync::Arc::from(&**id))
 }
